new version

+++ modules/cron.php
@@ -1,7 +1,11 @@
 <?php
 // Очищення тимчасових файлів
-if($_SERVER['REMOTE_ADDR'] == Core::$SERVER_IP){
+if($_SERVER['REMOTE_ADDR'] == '127.0.0.1'){
     function removeDirectory($dir){
+        if(!file_exists($dir)){
+            return false;
+        }
+
         if($objs = glob($dir."/*")){
             foreach($objs as $obj){
                 is_dir($obj)? removeDirectory($obj) : unlink($obj);
